requirements.txt
boto3==1.26.88
botocore==1.29.165
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
configparser==5.2.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h5py==3.9.0
keras==2.12.0
matplotlib==3.5.1
matplotlib-inline==0.1.6
numpy==1.23.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
psycopg2-binary==2.9.7
scikit-learn==1.3.1
scipy==1.11.2
tensorflow==2.12.0
tqdm==4.66.1
traitlets==5.14.1
ultralytics==8.0.188
